LoginScreen Update
User is now able to toggle between Login and Registration. Alerts (for registration) are replaced by an error text beneath the password field. In case of registration, there is now a second password field to repeat the password. Window is now scalable, because not "undecorated" anymore. Custom "X" deleted. Focus color is now orange (contrast color) instead of red.